[SwiftLint] Fixes remaining violations.
