+++ Changelog.md
@@ -1,6 +1,5 @@
 # Next
 
-
 - **Breaking Change** Removes `multipartBody` from TargetType protocol and adds a `task` instead.
 - **Breaking Change** Successful Response instances that have no data with them are now being converted to `.Success` `Result`s.
 - Adds Download and Upload Task type support to Moya.
