+++ Source/Moya+Defaults.swift
@@ -1,9 +1,7 @@
 import Alamofire
 
+/// These functions are default mappings to MoyaProvider's properties: endpoints, requests, manager, etc.
 public extension MoyaProvider {
-
-    // These functions are default mappings to MoyaProvider's properties: endpoints, requests, manager, etc.
-
     public final class func DefaultEndpointMapping(target: Target) -> Endpoint<Target> {
         let url = target.baseURL.URLByAppendingPathComponent(target.path).absoluteString
         return Endpoint(URL: url, sampleResponseClosure: {.NetworkResponse(200, target.sampleData)}, method: target.method, parameters: target.parameters)
